# source/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /user/update:
    put:
      tags:
        - user-ctrl
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /runner/update:
    put:
      tags:
        - runner-ctrl
      operationId: updateRunner
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Runner'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /run/update:
    put:
      tags:
        - run-ctrl
      operationId: updateRun
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Run'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /photo/update:
    put:
      tags:
        - photo-ctrl
      operationId: updatePhoto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Photo'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /article/update:
    put:
      tags:
        - article-ctrl
      operationId: updateArticle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /album/update:
    put:
      tags:
        - album-ctrl
      operationId: updateAlbum
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Album'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /user/reset:
    post:
      tags:
        - user-ctrl
      operationId: resetUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /user/login:
    post:
      tags:
        - user-ctrl
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /user/create:
    post:
      tags:
        - user-ctrl
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /runner/create:
    post:
      tags:
        - runner-ctrl
      operationId: createRunner
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Runner'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /run/create:
    post:
      tags:
        - run-ctrl
      operationId: createRun
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Run'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /photo/create:
    post:
      tags:
        - photo-ctrl
      operationId: createPhoto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Photo'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /article/create:
    post:
      tags:
        - article-ctrl
      operationId: createArticle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /album/create:
    post:
      tags:
        - album-ctrl
      operationId: createAlbum
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Album'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /user/read:
    get:
      tags:
        - user-ctrl
      operationId: readAllUser
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /user/read/{id}:
    get:
      tags:
        - user-ctrl
      operationId: readUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
  /user/check/{rstpwd}:
    get:
      tags:
        - user-ctrl
      operationId: loginUser_1
      parameters:
        - name: rstpwd
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /test:
    get:
      tags:
        - api
      operationId: testGet
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /runner/read:
    get:
      tags:
        - runner-ctrl
      operationId: readAllRunner
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Runner'
  /runner/readAllUser/{idRun}:
    get:
      tags:
        - runner-ctrl
      operationId: readAllUser_1
      parameters:
        - name: idRun
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Runner'
  /runner/readAllRun/{idUser}:
    get:
      tags:
        - runner-ctrl
      operationId: readAllRun
      parameters:
        - name: idUser
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Runner'
  /runner/read/{idUser}/{idRun}:
    get:
      tags:
        - runner-ctrl
      operationId: readRunner
      parameters:
        - name: idUser
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: idRun
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Runner'
  /run/read:
    get:
      tags:
        - run-ctrl
      operationId: readAllRun_1
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Run'
  /run/read/{id}:
    get:
      tags:
        - run-ctrl
      operationId: readRun
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Run'
  /photo/read:
    get:
      tags:
        - photo-ctrl
      operationId: readAllPhoto
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Photo'
  /photo/read/{id}:
    get:
      tags:
        - photo-ctrl
      operationId: readPhoto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Photo'
  /param/read:
    get:
      tags:
        - param-ctrl
      operationId: readParam
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Param'
  /param/delete/{id}:
    get:
      tags:
        - param-ctrl
      operationId: deleteParam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /param/create:
    get:
      tags:
        - param-ctrl
      operationId: createParam
      parameters:
        - name: key
          in: query
          required: false
          schema:
            type: string
            default: missing_key
        - name: value
          in: query
          required: false
          schema:
            type: string
            default: ''
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /article/read:
    get:
      tags:
        - article-ctrl
      operationId: readAllArticle
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
  /article/read/{id}:
    get:
      tags:
        - article-ctrl
      operationId: readArticle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Article'
  /album/read:
    get:
      tags:
        - album-ctrl
      operationId: readAllAlbum
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
  /album/read/{id}:
    get:
      tags:
        - album-ctrl
      operationId: readAlbum
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Album'
  /:
    get:
      tags:
        - api
      operationId: racineGet
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /user/delete/{id}:
    delete:
      tags:
        - user-ctrl
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /runner/delete/{idUser}/{idRun}:
    delete:
      tags:
        - runner-ctrl
      operationId: deleteRunner
      parameters:
        - name: idUser
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: idRun
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /run/delete/{id}:
    delete:
      tags:
        - run-ctrl
      operationId: deleteRun
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /photo/delete/{id}:
    delete:
      tags:
        - photo-ctrl
      operationId: deletePhoto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /article/delete/{id}:
    delete:
      tags:
        - article-ctrl
      operationId: deleteArticle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /album/delete/{id}:
    delete:
      tags:
        - album-ctrl
      operationId: deleteAlbum
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        firstName:
          type: string
        email:
          type: string
        password:
          type: string
        enable:
          type: boolean
        admin:
          type: boolean
    Runner:
      type: object
      properties:
        idUser:
          type: integer
          format: int32
        idRun:
          type: integer
          format: int32
        contribution:
          type: integer
          format: int32
        km:
          type: integer
          format: int32
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    Run:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        type:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    Photo:
      type: object
      properties:
        id:
          type: integer
          format: int32
        note:
          type: string
        file:
          type: string
        idAlbum:
          type: integer
          format: int32
    Article:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        note:
          type: string
        idPhoto:
          type: integer
          format: int32
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        news:
          type: boolean
        idUser:
          type: integer
          format: int32
        idRun:
          type: integer
          format: int32
    Album:
      type: object
      properties:
        id:
          type: integer
          format: int32
        note:
          type: string
        idUser:
          type: integer
          format: int32
        idRun:
          type: integer
          format: int32
    Login:
      type: object
      properties:
        id:
          type: string
        pwd:
          type: string
    Param:
      type: object
      properties:
        id:
          type: integer
          format: int32
        paramKey:
          type: string
        paramValue:
          type: string